Sub-class proteomics

This is a sub-class that is compatible to preprocessed data obtained from https://fragpipe.nesvilab.org/. It inherits all methods from the abstract class omics and only adapts the initialize function. It supports pre-existing data structures or paths to text files. When omics data is very large, data loading becomes very expensive. It is therefore recommended to use the reset() method to reset your changes. Every omics class creates an internal memory efficient back-up of the data, the resetting of changes is an instant process.

See also

Super class

omics -> proteomics

Active bindings

treeData

A "phylo" class, see as.phylo.

Methods

Public methods

proteomics$new()

proteomics$clone()

Inherited methods

omics$alpha_diversity()

omics$autoFlow()

omics$composition()

omics$copy()

omics$distance()

omics$feature_merge()

omics$feature_subset()

omics$foldchange()

omics$ordination()

omics$print()

omics$rankstat()

omics$removeNAs()

omics$reset()

omics$sample_subset()

omics$samplepair_subset()

omics$scale()

omics$validate()

proteomics$new()

Initializes the proteomics class object with proteomics$new()

Usage

proteomics$new(
  countData = NULL,
  metaData = NULL,
  featureData = NULL,
  treeData = NULL
)

Arguments

countData

A path to an existing file or a dense/sparse Matrix format.

metaData

A path to an existing file, data.table or data.frame.

featureData

A path to an existing file, data.table or data.frame.

treeData

A path to an existing newick file or class "phylo", see read.tree.

Returns

A new proteomics object.

proteomics$clone()

The objects of this class are cloneable with this method.

Usage

proteomics$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.